NIFI-1678: Started refactoring heartbeating mechanism, using a new package: org.apache.nifi.cluster.coordination

+++ nifi-nar-bundles/nifi-jms-bundle/nifi-jms-cf-service/src/test/java/org/apache/nifi/jms/cf/JMSConnectionFactoryProviderTest.java
@@ -69,9 +69,9 @@ public class JMSConnectionFactoryProviderTest {
         assertNotNull(cf);
         assertEquals("org.apache.nifi.jms.testcflib.TestConnectionFactory", cf.getClass().getName());
         assertEquals("myhost", this.get("getHost", cf));
-        assertEquals(1234, this.get("getPort", cf));
+        assertEquals(1234, (int) this.get("getPort", cf));
         assertEquals("foo", this.get("getFoo", cf));
-        assertEquals(3, this.get("getBar", cf));
+        assertEquals(3, (int) this.get("getBar", cf));
     }
 
     @Test(expected = AssertionError.class)
